Extract toInstantRange implicits & standardise naming
