+++ common/internal_model/src/main/scala/uk/ac/wellcome/models/parse/FuzzyDate.scala
@@ -1,8 +1,6 @@
 package uk.ac.wellcome.models.parse
 
-import scala.util.control.Exception
-import java.time.{DateTimeException, LocalDate}
-import uk.ac.wellcome.models.work.internal.InstantRange
+import java.time.LocalDate
 
 /**
   *  An exact or ambigous date
@@ -35,220 +33,6 @@ object FuzzyDateRange {
   ](from: FuzzyDateRange[FF, FT],
     to: FuzzyDateRange[TF, TT]): FuzzyDateRange[FF, TT] =
     FuzzyDateRange(from.from, to.to)
-}
-
-/**
-  *  Type class for conversion of TimePeriod types to InstantRange
-  */
-trait ToInstantRange[T <: TimePeriod] {
-
-  def apply(value: T): InstantRange
-
-  def safeConvert(value: T): Option[InstantRange] =
-    Exception.catching(classOf[DateTimeException]) opt apply(value)
-}
-
-object ToInstantRange extends DateHelpers {
-
-  def convert[T <: TimePeriod](f: T => InstantRange) = new ToInstantRange[T] {
-    def apply(value: T): InstantRange = f(value)
-  }
-
-  implicit val convertCalendarDate =
-    convert[CalendarDate](value =>
-      InstantRange(localDate(value), localDate(value), value.label))
-
-  implicit val convertYear =
-    convert[Year](value =>
-      InstantRange(yearStart(value.year), yearEnd(value.year), value.label))
-
-  implicit val convertMonthAndYear =
-    convert[MonthAndYear](
-      value =>
-        InstantRange(
-          monthStart(value.month, value.year),
-          monthEnd(value.month, value.year),
-          value.label))
-
-  implicit val convertCentury =
-    convert[Century](value =>
-      convertYearRange(centuryToFuzzyDateRange(value.century)))
-
-  implicit val convertCenturyRange =
-    convert[FuzzyDateRange[Century, Century]](
-      value =>
-        convertYearRange(
-          FuzzyDateRange.combine(
-            centuryToFuzzyDateRange(value.from.century),
-            centuryToFuzzyDateRange(value.to.century)
-          )
-      )
-    )
-
-  implicit val convertCenturyAndDecade =
-    convert[CenturyAndDecade](
-      value =>
-        convertYearRange(
-          centuryAndDecadeToFuzzyDateRange(value.century, value.decade)))
-
-  implicit val convertDecadeRange =
-    convert[FuzzyDateRange[CenturyAndDecade, CenturyAndDecade]](
-      value =>
-        convertYearRange(
-          FuzzyDateRange.combine(
-            centuryAndDecadeToFuzzyDateRange(
-              value.from.century,
-              value.from.decade),
-            centuryAndDecadeToFuzzyDateRange(value.to.century, value.to.decade)
-          )
-      )
-    )
-
-  implicit val convertYearCentury =
-    convert[FuzzyDateRange[Year, Century]](
-      value =>
-        convertYearRange(
-          FuzzyDateRange(
-            value.from,
-            centuryToFuzzyDateRange(value.to.century).to
-          )
-      )
-    )
-
-  implicit val convertCenturyYear =
-    convert[FuzzyDateRange[Century, Year]](
-      value =>
-        convertYearRange(
-          FuzzyDateRange(
-            centuryToFuzzyDateRange(value.from.century).from,
-            value.to
-          )
-      )
-    )
-
-  implicit val convertYearRange =
-    convert[FuzzyDateRange[Year, Year]](
-      value =>
-        InstantRange(
-          yearStart(value.from.year),
-          yearEnd(value.to.year),
-          value.label))
-
-  implicit val convertCalendarDateRange =
-    convert[FuzzyDateRange[CalendarDate, CalendarDate]](value =>
-      InstantRange(localDate(value.from), localDate(value.to), value.label))
-
-  implicit val convertCalendarMonth =
-    convert[FuzzyDateRange[CalendarDate, MonthAndYear]](
-      value =>
-        InstantRange(
-          localDate(value.from),
-          monthEnd(value.to.month, value.to.year),
-          value.label))
-
-  implicit val convertMonthCalendar =
-    convert[FuzzyDateRange[MonthAndYear, CalendarDate]](
-      value =>
-        InstantRange(
-          monthStart(value.from.month, value.from.year),
-          localDate(value.to),
-          value.label
-      ))
-
-  implicit val convertMonthToDateWithinYear =
-    convert[FuzzyDateRange[MonthAndYear, MonthAndDay]](
-      value =>
-        InstantRange(
-          monthStart(value.from.month, value.from.year),
-          localDate(value.to.day, value.to.month, value.from.year),
-          value.label))
-
-  implicit val convertCalendarDateToMonthDateWithinYear =
-    convert[FuzzyDateRange[CalendarDate, MonthAndDay]](
-      value =>
-        InstantRange(
-          localDate(value.from),
-          localDate(value.to.day, value.to.month, value.from.year),
-          value.label))
-
-  implicit val convertMonthDateToCalendarDateWithinYear =
-    convert[FuzzyDateRange[MonthAndDay, CalendarDate]](
-      value =>
-        InstantRange(
-          localDate(value.from.day, value.from.month, value.to.year),
-          localDate(value.to),
-          value.label))
-
-  implicit val convertMonthRangeAcrossYears =
-    convert[FuzzyDateRange[MonthAndYear, MonthAndYear]](
-      value =>
-        InstantRange(
-          monthStart(value.from.month, value.from.year),
-          monthEnd(value.to.month, value.to.year),
-          value.label))
-
-  implicit val convertYearToMonth =
-    convert[FuzzyDateRange[Year, MonthAndYear]](
-      value =>
-        InstantRange(
-          yearStart(value.from.year),
-          monthEnd(value.to.month, value.to.year),
-          value.label))
-
-  implicit val convertYearToDate =
-    convert[FuzzyDateRange[Year, CalendarDate]](
-      value =>
-        InstantRange(
-          yearStart(value.from.year),
-          localDate(value.to),
-          value.label))
-
-  implicit val convertDateToYear =
-    convert[FuzzyDateRange[CalendarDate, Year]](value =>
-      InstantRange(localDate(value.from), yearEnd(value.to.year), value.label))
-
-  implicit val convertMonthRangeWithinAYear =
-    convert[FuzzyDateRange[Month, MonthAndYear]](
-      value =>
-        InstantRange(
-          monthStart(value.from.month, value.to.year),
-          monthEnd(value.to.month, value.to.year),
-          value.label))
-
-  implicit val convertDayToMonthWithinYear =
-    convert[FuzzyDateRange[CalendarDate, Month]](
-      value =>
-        InstantRange(
-          localDate(value.from),
-          monthEnd(value.to.month, value.from.year),
-          value.label
-      )
-    )
-
-  implicit val convertMonthRangeWithinYearReversed =
-    convert[FuzzyDateRange[MonthAndYear, Month]](
-      value =>
-        InstantRange(
-          monthStart(value.from.month, value.from.year),
-          monthEnd(value.to.month, value.from.year),
-          value.label
-      ))
-
-  implicit val convertDayRangeWithinAMonth =
-    convert[FuzzyDateRange[Day, CalendarDate]](
-      value =>
-        InstantRange(
-          localDate(value.from.day, value.to.month, value.to.year),
-          localDate(value.to),
-          value.label))
-
-  implicit val convertDayRangeWithinAMonthReversed =
-    convert[FuzzyDateRange[CalendarDate, Day]](
-      value =>
-        InstantRange(
-          localDate(value.from),
-          localDate(value.to.day, value.from.month, value.from.year),
-          value.label))
 }
 
 sealed trait TimePeriod extends DateHelpers {
